Enhance ONNX function calling with improved parsing and invocation mechanisms
- Updated `OnnxFunctionCalling` demo to include a more detailed system message for the assistant, allowing for better context and reasoning in responses.
- Introduced `OnnxFunctionCallParser` to handle function call parsing from model responses, improving robustness and clarity.
- Added `OnnxFunctionInvoker` to manage function invocation and natural language response generation, streamlining the process.
- Implemented `OnnxResponseFormatter` for better formatting of function results, enhancing user experience.
- Refactored `OnnxRuntimeGenAIFunctionCallingChatCompletionService` to utilize the new parser and invoker, improving code organization and maintainability.

These changes significantly enhance the functionality and usability of the ONNX function calling capabilities, providing a more intuitive interaction model.
